Improve: Hide more APIs that shouldn't be public in file-formats-api.
Also move embeddable-transform-details to formats folder, where it belongs.
