+++ projects/api/dxt-lossless-transform-file-formats-api/src/embed/formats/bc3.rs
@@ -1,6 +1,7 @@
 //! BC3 format file format support.
 
-use crate::embed::{EmbedError, EmbeddableTransformDetails, TransformFormat};
+use super::EmbeddableTransformDetails;
+use crate::embed::{EmbedError, TransformFormat};
 use bitfield::bitfield;
 
 /// Header version for BC3 format
