+++ projects/api/dxt-lossless-transform-file-formats-api/src/embed/formats/mod.rs
@@ -1,11 +1,36 @@
 //! BCx format-specific embeddable implementations.
 
-pub mod bc1;
-pub mod bc2;
-pub mod bc3;
-pub mod bc7;
+pub(crate) mod bc1;
+#[allow(dead_code)]
+pub(crate) mod bc2; // code not ready, placeholder
+#[allow(dead_code)]
+pub(crate) mod bc3; // code not ready, placeholder
+#[allow(dead_code)]
+pub(crate) mod bc7; // code not ready, placeholder
 
-pub use bc1::EmbeddableBc1Details;
-pub use bc2::EmbeddableBc2Details;
-pub use bc3::EmbeddableBc3Details;
-pub use bc7::EmbeddableBc7Details;
+pub(crate) use bc1::EmbeddableBc1Details;
+
+use super::{EmbedError, TransformFormat, TransformHeader};
+
+/// Trait for transform details that can be embedded in a 4-byte header.
+///
+/// Each BCx format implements this trait to define how its transform details
+/// are packed into the 28 bits of format-specific data in the header.
+pub(crate) trait EmbeddableTransformDetails: Sized {
+    /// The transform format this implementation is for
+    const FORMAT: TransformFormat;
+
+    /// Pack the transform details into a 28-bit value
+    fn pack(&self) -> u32;
+
+    /// Unpack transform details from a 28-bit value
+    fn unpack(data: u32) -> Result<Self, EmbedError>;
+
+    /// Extract from a complete transform header
+    fn from_header(header: TransformHeader) -> Result<Self, EmbedError> {
+        if header.format() != Self::FORMAT {
+            return Err(EmbedError::FormatMismatch);
+        }
+        Self::unpack(header.format_data())
+    }
+}
